api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["al-ov73 <klever1987@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
minio = "^7.2.7"
pytest = "^8.3.1"
httpx = "^0.27.0"
trio = "^0.26.0"
gunicorn = "^22.0.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiohttp = "^3.10.5"
redis = "^5.0.8"
fastapi-cache2 = "^0.2.2"
aioredis = "^2.0.1"
websockets = "^13.0"
sqlalchemy-utils = "^0.41.2"
flake8 = "^7.1.1"
pytest-cov = "^5.0.0"
gigachat = "^0.1.35"
telethon = "^1.38.1"
aiogram = "^3.15.0"
apscheduler = "^3.11.0"
vk-api = "^11.9.9"
filelock = "^3.16.1"
pillow = "^11.1.0"
sqladmin = "^0.20.1"
itsdangerous = "^2.2.0"
openai = "^1.61.0"
pytz = "^2025.1"
tinydb = "^4.8.2"
jinja2 = "^3.1.6"
pdf2image = "^1.17.0"
pypdf2 = "^3.0.1"
pymupdf = "^1.25.5"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
